Add messaging between gateway and enforcement
Add a script to read ip neighbour and output the important bits in
json. The idea here is that said json would be sent to the enforcement
device, which would use it to determine if it needed to change its
state. For now, we just write to a file to represent that pipe.

Also update the README to mention a new dependency.